backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2  # MongoDB async driver
pydantic==2.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
python-multipart==0.0.6
cors==1.0.1